move frontier backend to kv namespace
