+++ client/cli/src/frontier_db_cmd/mod.rs
@@ -99,7 +99,7 @@ impl FrontierDbCmd {
 	pub fn run<C, B: BlockT>(
 		&self,
 		client: Arc<C>,
-		backend: Arc<fc_db::Backend<B>>,
+		backend: Arc<fc_db::kv::Backend<B>>,
 	) -> sc_cli::Result<()>
 	where
 		C: sp_api::ProvideRuntimeApi<B>,
