+++ client/db/src/kv/parity_db_adapter.rs
@@ -18,7 +18,7 @@
 
 use sp_database::{error::DatabaseError, Change, ColumnId, Transaction};
 
-use crate::Database;
+use crate::kv::Database;
 
 fn handle_err<T>(result: parity_db::Result<T>) -> T {
 	match result {
@@ -62,6 +62,6 @@ impl<H: Clone + AsRef<[u8]>> Database<H> for DbAdapter {
 	}
 
 	fn sanitize_key(&self, key: &mut Vec<u8>) {
-		let _prefix = key.drain(0..key.len() - crate::DB_HASH_LEN);
+		let _prefix = key.drain(0..key.len() - crate::kv::DB_HASH_LEN);
 	}
 }
