+++ template/node/src/command.rs
@@ -20,7 +20,7 @@ use clap::Parser;
 use sc_cli::{ChainSpec, RuntimeVersion, SubstrateCli};
 use sc_service::DatabaseSource;
 // Frontier
-use fc_db::frontier_database_dir;
+use fc_db::kv::frontier_database_dir;
 
 use crate::{
 	chain_spec,
